+++ api/src/pdc/services/operator/OperatorServiceProvider.ts
@@ -2,14 +2,6 @@ import { ExtensionInterface, NewableType, serviceProvider } from "@/ilos/common/
 import { ServiceProvider as AbstractServiceProvider } from "@/ilos/core/index.ts";
 import { defaultMiddlewareBindings } from "@/pdc/providers/middleware/index.ts";
 import { ValidatorExtension, ValidatorMiddleware } from "@/pdc/providers/validator/index.ts";
-import { binding as createBinding } from "@/shared/operator/create.schema.ts";
-import { binding as deleteBinding } from "@/shared/operator/delete.schema.ts";
-import { binding as findBinding } from "@/shared/operator/find.schema.ts";
-import { binding as findbyuuidBinding } from "@/shared/operator/findbyuuid.schema.ts";
-import { binding as patchContactsBinding } from "@/shared/operator/patchContacts.schema.ts";
-import { binding as patchThumbnailBinding } from "@/shared/operator/patchThumbnail.schema.ts";
-import { binding as quickfindBinding } from "@/shared/operator/quickfind.schema.ts";
-import { binding as updateBinding } from "@/shared/operator/update.schema.ts";
 import { CreateOperatorAction } from "./actions/CreateOperatorAction.ts";
 import { DeleteOperatorAction } from "./actions/DeleteOperatorAction.ts";
 import { FindByUuidOperatorAction } from "./actions/FindByUuidOperatorAction.ts";
@@ -20,6 +12,14 @@ import { PatchThumbnailOperatorAction } from "./actions/PatchThumbnailOperatorAc
 import { QuickFindOperatorAction } from "./actions/QuickFindOperatorAction.ts";
 import { UpdateOperatorAction } from "./actions/UpdateOperatorAction.ts";
 import { config } from "./config/index.ts";
+import { binding as createBinding } from "./contracts/create.schema.ts";
+import { binding as deleteBinding } from "./contracts/delete.schema.ts";
+import { binding as findBinding } from "./contracts/find.schema.ts";
+import { binding as findbyuuidBinding } from "./contracts/findbyuuid.schema.ts";
+import { binding as patchContactsBinding } from "./contracts/patchContacts.schema.ts";
+import { binding as patchThumbnailBinding } from "./contracts/patchThumbnail.schema.ts";
+import { binding as quickfindBinding } from "./contracts/quickfind.schema.ts";
+import { binding as updateBinding } from "./contracts/update.schema.ts";
 import { OperatorPgRepositoryProvider } from "./providers/OperatorPgRepositoryProvider.ts";
 
 @serviceProvider({
