+++ api/src/pdc/services/operator/actions/ListOperatorAction.ts
@@ -1,16 +1,9 @@
 import { handler } from "@/ilos/common/index.ts";
 import { Action as AbstractAction } from "@/ilos/core/index.ts";
-import {
-  contentBlacklistMiddleware,
-  hasPermissionMiddleware,
-} from "@/pdc/providers/middleware/index.ts";
+import { contentBlacklistMiddleware, hasPermissionMiddleware } from "@/pdc/providers/middleware/index.ts";
 
+import { handlerConfig, ParamsInterface, ResultInterface } from "../contracts/list.contract.ts";
 import { OperatorRepositoryProviderInterfaceResolver } from "../interfaces/OperatorRepositoryProviderInterface.ts";
-import {
-  handlerConfig,
-  ParamsInterface,
-  ResultInterface,
-} from "@/shared/operator/list.contract.ts";
 
 @handler({
   ...handlerConfig,
