+++ api/src/pdc/services/operator/actions/PatchThumbnailOperatorAction.ts
@@ -1,17 +1,10 @@
 import { handler } from "@/ilos/common/index.ts";
 import { Action as AbstractAction } from "@/ilos/core/index.ts";
-import {
-  copyFromContextMiddleware,
-  hasPermissionByScopeMiddleware,
-} from "@/pdc/providers/middleware/index.ts";
+import { copyFromContextMiddleware, hasPermissionByScopeMiddleware } from "@/pdc/providers/middleware/index.ts";
 
+import { handlerConfig, ParamsInterface, ResultInterface } from "../contracts/patchThumbnail.contract.ts";
+import { alias } from "../contracts/patchThumbnail.schema.ts";
 import { OperatorRepositoryProviderInterfaceResolver } from "../interfaces/OperatorRepositoryProviderInterface.ts";
-import {
-  handlerConfig,
-  ParamsInterface,
-  ResultInterface,
-} from "@/shared/operator/patchThumbnail.contract.ts";
-import { alias } from "@/shared/operator/patchThumbnail.schema.ts";
 
 @handler({
   ...handlerConfig,
