+++ api/src/pdc/services/operator/actions/UpdateOperatorAction.ts
@@ -1,18 +1,11 @@
 import { handler } from "@/ilos/common/index.ts";
 import { Action as AbstractAction } from "@/ilos/core/index.ts";
-import {
-  copyFromContextMiddleware,
-  hasPermissionByScopeMiddleware,
-} from "@/pdc/providers/middleware/index.ts";
+import { copyFromContextMiddleware, hasPermissionByScopeMiddleware } from "@/pdc/providers/middleware/index.ts";
 
+import { handlerConfig, ParamsInterface, ResultInterface } from "../contracts/update.contract.ts";
+import { alias } from "../contracts/update.schema.ts";
 import { phoneComplianceHelper } from "../helpers/phoneComplianceHelper.ts";
 import { OperatorRepositoryProviderInterfaceResolver } from "../interfaces/OperatorRepositoryProviderInterface.ts";
-import {
-  handlerConfig,
-  ParamsInterface,
-  ResultInterface,
-} from "@/shared/operator/update.contract.ts";
-import { alias } from "@/shared/operator/update.schema.ts";
 
 @handler({
   ...handlerConfig,
