+++ api/src/pdc/services/operator/contracts/common/interfaces/OperatorDbInterface.ts
