+++ api/src/pdc/services/operator/contracts/delete.contract.ts
