+++ api/src/pdc/services/operator/contracts/delete.schema.ts
