+++ api/src/pdc/services/operator/contracts/find.schema.ts
