+++ api/src/pdc/services/operator/contracts/findbyuuid.contract.ts
