+++ api/src/pdc/services/operator/contracts/findbyuuid.schema.ts
