+++ api/src/pdc/services/operator/contracts/list.contract.ts
