+++ api/src/pdc/services/operator/contracts/patch.contract.ts
