+++ api/src/pdc/services/operator/contracts/patchContacts.schema.ts
@@ -1,0 +1,14 @@
+import { contacts } from "../../../../shared/common/schemas/contacts.ts";
+
+export const alias = "operator.patchContacts";
+export const schema = {
+  $id: alias,
+  type: "object",
+  required: ["_id", "patch"],
+  additionalProperties: false,
+  properties: {
+    _id: { macro: "serial" },
+    patch: contacts,
+  },
+};
+export const binding = [alias, schema];
