+++ api/src/pdc/services/operator/contracts/patchThumbnail.contract.ts
