+++ api/src/pdc/services/operator/contracts/patchThumbnail.schema.ts
